can now sign in with Google or Discord through NextAuth
